Add the ability to specify additional logging functions
Adds the ability to specify additional logging functions beyond
the svclog.Logger functions that should be checked to ensure that
a parameter at a particular position is a compile-time constant
using configuration.

This allows for more safetly supporting the pattern where a project
may define a custom logging function that wraps an svc1log logger:
in this scenario, the svc1log call in the wrapper function can be
marked as allowed and then the wrapper function can be added to the
check to ensure that code that calls the wrapper function uses
compile-time constants. Also allows for performing this check on
arbitrary functions that are not directly related to svc1log.